Add backButton functionality + cleanup committed
The backButton now check if we are inside of an operation using the
isOperationInPlace boolean variable.
If the value is true it means that we have on screen the result
of a computation and we don't want to escape any character from that.

Code cleanup has been done on comments and unused variables